src/requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
gevent==20.12.1
gevent-websocket==0.10.1
graphviz==0.16
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
memory-profiler==0.58.0
objgraph==3.5.0
psutil==5.8.0
pymongo==3.11.2
python-decouple==3.4
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2020.5
redis==3.5.3
requests==2.25.1
rq==1.7.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
tzlocal==2.1
urllib3==1.26.2
uWSGI==2.0.19.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
